Dynamic parameters (#16)
Allows a node to contain an object capable of handling dynamic parameters.
Contains methods for user defined parameter validation callbacks and
notifies the user if the parameters changed.

Run the example:
```
ros2 run rosparam_shortcuts dynamic_parameters_example
```

The parameter `length` is static and will not change if requested. The
parameter `damping` is valid for any range of values. And the parameter
`mass` is only valid on the domain (0 10).

Try changing a parameters:
```
ros2 param set /dynamic_parameters damping 100.0
```